+++ CHANGELOG.md
@@ -7,10 +7,13 @@
 - Able to auto upgrade or manual check for upgrade
 
 ### Fixes
+- Fix mac icon issue
 
 ### Improvements
-
+- Better error handling using ErrorBoundary, allow user self recover from error state
 - Code rebranding, replace `Odin` in code
+- Add sentry integration
+- Customized sentry feedback form
 
 # v0.9.2 (2019.04.09)